raw conversion of UnconfiguredFake specs
